# apps/server/auth-service/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/mfa/enroll/start": {
      "post": {
        "operationId": "MfaLoginController_enrollStart",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaEnrollStartDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MfaEnrollStartResponseDto"
                }
              }
            }
          }
        },
        "summary": "Start authenticator enrollment (QR + manual key)",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/mfa/enroll/confirm": {
      "post": {
        "operationId": "MfaLoginController_enrollConfirm",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaEnrollConfirmDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MfaEnrollConfirmResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "MFA_TOKEN_INVALID / MFA_CODE_INVALID / MFA_TOO_MANY_ATTEMPTS"
          }
        },
        "summary": "Confirm enrollment with one code; returns the backup codes, NO session yet",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/users/me/mfa/backup-codes": {
      "post": {
        "operationId": "MfaSelfController_regenerateBackupCodes",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegenerateBackupCodesDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupCodesResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Replace the backup codes (needs the current authenticator code)",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/me/mfa/enroll/start": {
      "post": {
        "operationId": "MfaSelfController_enrollStart",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MfaEnrollStartResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "MFA_ALREADY_ENROLLED"
          },
          "403": {
            "description": "SSO_REQUIRED: the workspace requires SSO for this member (no PON 2FA)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Turn 2FA on from Settings: QR + manual key of a pending secret (10 minutes, the same one on every call)",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/me/mfa/enroll/confirm": {
      "post": {
        "operationId": "MfaSelfController_enrollConfirm",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaSelfEnrollConfirmDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MfaSelfEnrollConfirmResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "MFA_CODE_INVALID (params.remaining) / MFA_TOO_MANY_ATTEMPTS / MFA_NOT_ENROLLED (no pending secret: start again) / MFA_ALREADY_ENROLLED"
          },
          "403": {
            "description": "SSO_REQUIRED: the workspace requires SSO for this member (no PON 2FA)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Confirm the Settings enrollment with one code; returns the backup codes once (session unchanged)",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/me/mfa/disable": {
      "post": {
        "operationId": "MfaSelfController_disable",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaDisableDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "MFA_REQUIRED_BY_ROLE (Owner / Admin-like) / MFA_NOT_ENROLLED / MFA_CODE_INVALID (params.remaining) / MFA_TOO_MANY_ATTEMPTS"
          },
          "403": {
            "description": "SSO_REQUIRED: the workspace requires SSO for this member (no PON 2FA)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Turn optional 2FA off (a current authenticator code or an unused backup code)",
        "tags": [
          "users"
        ]
      }
    },
    "/admin/members/{id}/mfa/reset": {
      "post": {
        "operationId": "MfaAdminController_reset",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reset a member's 2FA and revoke their sessions (Owner: anyone else; MANAGE_MEMBERS: non-admin members)",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Platform Auth Service",
    "description": "Authentication, OTP, refresh tokens, user search and friends API.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "MfaEnrollStartDto": {
        "type": "object",
        "properties": {
          "mfaToken": {
            "type": "string",
            "description": "mfaToken from the MFA_REQUIRED sign-in answer"
          }
        },
        "required": [
          "mfaToken"
        ]
      },
      "MfaEnrollStartResponseDto": {
        "type": "object",
        "properties": {
          "otpauthUrl": {
            "type": "string",
            "example": "otpauth://totp/PON:jane%40acme.com?secret=...&issuer=PON"
          },
          "secret": {
            "type": "string",
            "description": "Base32 secret for manual entry"
          },
          "qrDataUrl": {
            "type": "string",
            "description": "QR code of otpauthUrl as a data:image/png;base64 URL"
          }
        },
        "required": [
          "otpauthUrl",
          "secret",
          "qrDataUrl"
        ]
      },
      "MfaEnrollConfirmDto": {
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "string",
            "description": "Device identifier for the issued session"
          },
          "platform": {
            "type": "string",
            "description": "Originating platform",
            "example": "web"
          },
          "mfaToken": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "example": "123456",
            "description": "6-digit code from the authenticator app"
          }
        },
        "required": [
          "mfaToken",
          "code"
        ]
      },
      "MfaEnrollConfirmResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "MFA_BACKUP_CODES_ISSUED"
            ],
            "example": "MFA_BACKUP_CODES_ISSUED"
          },
          "backupCodes": {
            "example": [
              "ABCDE-FGH23"
            ],
            "description": "10 single-use backup codes (stored hashed). Re-readable via enroll/codes until enroll/complete, at most 10 minutes.",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "code",
          "backupCodes"
        ]
      },
...
      "SuccessResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "success"
        ]
      },
      "RegenerateBackupCodesDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "123456",
            "description": "Current authenticator code"
          }
        },
        "required": [
          "code"
        ]
      },
      "BackupCodesResponseDto": {
        "type": "object",
        "properties": {
          "backupCodes": {
            "description": "10 new single-use backup codes (the old ones stop working). Shown once.",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "backupCodes"
        ]
      },
      "MfaSelfEnrollConfirmDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "123456",
            "description": "6-digit code from the authenticator app just set up"
          }
        },
        "required": [
          "code"
        ]
      },
      "MfaSelfEnrollConfirmResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "MFA_BACKUP_CODES_ISSUED"
            ],
            "example": "MFA_BACKUP_CODES_ISSUED"
          },
          "backupCodes": {
            "example": [
              "ABCDE-FGH23"
            ],
            "description": "10 single-use backup codes (stored hashed). Shown once: they cannot be read again.",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "code",
          "backupCodes"
        ]
      },
      "MfaDisableDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "123456",
            "description": "Current authenticator code (send exactly one of code / backupCode)"
          },
          "backupCode": {
            "type": "string",
            "example": "ABCDE-FGH23",
            "description": "An unused backup code (send exactly one of code / backupCode)"
          }
        }
      },
...
    }
  }
}